+++ dashboard/src/constants.ts
@@ -39,10 +39,16 @@ export const ProtocolTypeEntityNames: Record<string, string> = {
 export interface Schema {
   entities: string[];
   entitiesData: { [x: string]: { [x: string]: string } };
-  query: string;
   poolData: { [x: string]: string };
   events: string[];
   protocolFields: { [x: string]: string };
+  query: string;
+  financialsQuery: string;
+  hourlyUsageQuery: string;
+  dailyUsageQuery: string;
+  protocolTableQuery: string;
+  poolsQuery: string;
+  poolTimeseriesQuery: string;
 }
 export const percentageFieldList = [
   "rates",
@@ -59,9 +65,9 @@ export const percentageFieldList = [
 // negativeFieldList contains field names that can be negative
 export const negativeFieldList = [];
 export const ProtocolsToQuery: { [name: string]: { [network: string]: string } } = {
-  aaveV2: {
-    mainnet: "https://api.thegraph.com/subgraphs/name/messari/aave-v2-ethereum",
-  },
+  // aaveV2: {
+  //   mainnet: "https://api.thegraph.com/subgraphs/name/messari/aave-v2-ethereum",
+  // },
   abracadabra: {
     mainnet: "https://api.thegraph.com/subgraphs/name/messari/abracadabra-money-ethereum",
     avalanche: "https://api.thegraph.com/subgraphs/name/messari/abracadabra-money-avalanche",
