+++ dashboard/src/interfaces/AllDataTabs.tsx
@@ -1,5 +1,5 @@
 import { TabContext, TabPanel } from "@mui/lab";
-import { Tab, Tabs } from "@mui/material";
+import { CircularProgress, Tab, Tabs } from "@mui/material";
 import React from "react";
 import { ProtocolTypeEntityName, ProtocolTypeEntityNames } from "../constants";
 import EventsTab from "./tabs/EventsTab";
@@ -16,7 +16,6 @@ const StyledTabs = styled(Tabs)`
 
 interface AllDataTabsProps {
   data: any;
-  entities: string[];
   entitiesData: { [x: string]: { [x: string]: string } };
   protocolFields: { [x: string]: string };
   tabValue: string;
@@ -32,12 +31,19 @@ interface AllDataTabsProps {
   paginate: React.Dispatch<React.SetStateAction<number>>;
   skipAmt: number;
   poolOverviewRequest: { [x: string]: any };
+  poolTimeseriesRequest: { [x: string]: any };
+  protocolTimeseriesData: any;
+  protocolTimeseriesLoading: any;
+  protocolTimeseriesError: any;
+  protocolTableData: any;
+  poolsListData: { [x: string]: any };
+  poolListLoading: any;
+  poolsListError: any;
 }
 
 // This component is for each individual subgraph
 function AllDataTabs({
   data,
-  entities,
   entitiesData,
   protocolFields,
   tabValue,
@@ -53,25 +59,34 @@ function AllDataTabs({
   setProtocolId,
   skipAmt,
   poolOverviewRequest,
+  poolTimeseriesRequest,
+  protocolTimeseriesData,
+  protocolTableData,
+  poolsListData,
+  poolListLoading,
+  protocolTimeseriesLoading,
+  protocolTimeseriesError,
+  poolsListError,
 }: AllDataTabsProps) {
   const protocolEntityName = ProtocolTypeEntityName[data.protocols[0].type];
   const protocolEntityNames = ProtocolTypeEntityNames[data.protocols[0].type];
 
   let protocolDropDown = null;
-  if (data[protocolEntityNames].length > 1) {
+  if (data.protocols.length > 1) {
     protocolDropDown = (
       <div style={{ padding: "24px" }}>
-        <ProtocolDropDown setProtocolId={(x) => setProtocolId(x)} protocols={data[protocolEntityNames]} />
+        <ProtocolDropDown setProtocolId={(x) => setProtocolId(x)} protocols={data.protocols} />
       </div>
     );
   }
-  let protocolData = data[protocolEntityName];
-  if (!protocolData) {
-    protocolData = data[protocolEntityNames][0];
-  }
-  if (protocolData?.lendingType === "CDP") {
+
+  if (protocolTableData?.lendingType === "CDP") {
     protocolFields.mintedTokens += "!";
     protocolFields.mintedTokenSupplies += "!";
+  }
+
+  if (!protocolTableData) {
+    return <CircularProgress sx={{ margin: 6 }} size={50} />;
   }
 
   return (
@@ -79,54 +94,64 @@ function AllDataTabs({
       <TabContext value={tabValue}>
         <StyledTabs value={tabValue} onChange={handleTabChange}>
           <Tab label="Protocol" value="1" />
-          <Tab label="Pool" value="2" />
-          <Tab label="Events" value="3" />
-          <Tab label="Pool Overview" value="4" />
+          <Tab label="Pool Overview" value="2" />
+          <Tab label="Pool" value="3" />
+          <Tab label="Events" value="4" />
         </StyledTabs>
         {protocolDropDown}
         <TabPanel value="1">
           {/* PROTOCOL TAB */}
           <ProtocolTab
-            data={data}
-            protocolData={protocolData}
-            entities={entities}
+            protocolType={data.protocols[0].type}
+            protocolTableData={protocolTableData}
             entitiesData={entitiesData}
             protocolFields={protocolFields}
+            protocolTimeseriesData={protocolTimeseriesData}
+            protocolTimeseriesLoading={protocolTimeseriesLoading}
+            protocolTimeseriesError={protocolTimeseriesError}
           />
         </TabPanel>
         <TabPanel value="2">
+          {/* POOLOVERVIEW TAB */}
+          <PoolOverviewTab
+            pools={pools}
+            subgraphToQueryURL={subgraphToQueryURL}
+            protocolType={data.protocols[0].type}
+            poolOverviewRequest={poolOverviewRequest}
+            setPoolId={(x) => setPoolId(x)}
+            paginate={(x) => paginate(x)}
+            skipAmt={skipAmt}
+            handleTabChange={(x, y) => handleTabChange(x, y)}
+          />
+        </TabPanel>
+        <TabPanel value="3">
           {/* POOL TAB */}
           <PoolTab
             data={data}
-            entities={entities}
+            poolTimeseriesData={poolTimeseriesRequest.poolTimeseriesData}
+            poolTimeseriesLoading={poolTimeseriesRequest.poolTimeseriesLoading}
+            poolTimeseriesError={poolTimeseriesRequest.poolTimeseriesError}
             entitiesData={entitiesData}
             poolId={poolId}
             setPoolId={(x) => setPoolId(x)}
             poolData={poolData}
-            protocolData={protocolData}
+            protocolData={protocolTableData}
+            poolsListError={poolsListError}
+            poolsList={poolsListData}
+            poolListLoading={poolListLoading}
+            poolNames={poolNames}
           />
         </TabPanel>
-        <TabPanel value="3">
+        <TabPanel value="4">
           {/* EVENTS TAB */}
           <EventsTab
             data={data}
             events={events}
             poolId={poolId}
             setPoolId={(x) => setPoolId(x)}
+            poolsList={poolsListData}
+            poolListLoading={poolListLoading}
             poolNames={poolNames}
-          />
-        </TabPanel>
-        <TabPanel value="4">
-          {/* POOLOVERVIEW TAB */}
-          <PoolOverviewTab
-            pools={pools}
-            subgraphToQueryURL={subgraphToQueryURL}
-            protocolData={protocolData}
-            poolOverviewRequest={poolOverviewRequest}
-            setPoolId={(x) => setPoolId(x)}
-            paginate={(x) => paginate(x)}
-            skipAmt={skipAmt}
-            handleTabChange={(x, y) => handleTabChange(x, y)}
           />
         </TabPanel>
       </TabContext>
