+++ dashboard/src/interfaces/IssuesDisplay.tsx
@@ -1,5 +1,5 @@
 import { styled } from "../styled";
-import { Typography } from "@mui/material";
+import { CircularProgress, Typography } from "@mui/material";
 import React, { useEffect } from "react";
 import { useState } from "react";
 
@@ -28,7 +28,7 @@ const messagesByLevel = (
     for (let x = 0; x < issuesArray.length; x++) {
       let issuesMsg = (issuesArray[x].fieldName ? issuesArray[x].fieldName + ": " : "") + issuesArray[x].message;
       if (issuesArray[x].type === "SUM") {
-        issuesMsg = `All values in ${issuesArray[x].fieldName} are zero. Verify that this data is being mapped correctly.`;
+        issuesMsg = `All values in ${issuesArray[x].fieldName} are 0. Verify that this data is being mapped correctly.`;
       }
       if (issuesArray[x].type === "LIQ") {
         issuesMsg = `${issuesArray[x].fieldName} timeseries value cannot be higher than totalValueLockedUSD on the pool. Look at snapshot id ${issuesArray[x].message}`;
@@ -38,11 +38,10 @@ const messagesByLevel = (
           ${issuesArray[x].fieldName} cumulative value dropped on snapshot id ${issuesArray[x].message}. Cumulative values should always increase.`;
       }
       if (issuesArray[x].type === "TVL-") {
-        issuesMsg = `totalValueLockedUSD on ${issuesArray[x].fieldName} is below 1000.`;
+        issuesMsg = `${issuesArray[x].fieldName} is below 1000.`;
       }
       if (issuesArray[x].type === "TVL+") {
-        issuesMsg = `
-          totalValueLockedUSD on ${issuesArray[x].fieldName} is above 1,000,000,000,000.`;
+        issuesMsg = `${issuesArray[x].fieldName} is above 1,000,000,000,000.`;
       }
       if (issuesArray[x].type === "DEC") {
         issuesMsg = `Decimals on ${issuesArray[x].fieldName} could not be pulled. The default decimal value of 18 has been applied.`;
@@ -54,14 +53,14 @@ const messagesByLevel = (
         issuesMsg = issuesArray[x].message;
       }
       if (issuesArray[x].type === "TOK") {
-        issuesMsg = `'${issuesArray[x].fieldName}' in the timeseries data refers to a token that does not exist on this pool. '${issuesArray[x].message}' is an invalid index.`;
+        issuesMsg = `'${issuesArray[x].fieldName}' in the timeseries data refers to a token that does not exist on this pool. '${issuesArray[x].message}' references an invalid index.`;
       }
       if (issuesArray[x].type === "NEG") {
         const msgObj = JSON.parse(issuesArray[x].message);
         issuesMsg = `'${issuesArray[x].fieldName}' has ${msgObj.count} negative values. First instance of a negative value is on snapshot ${msgObj.firstSnapshot} with a value of ${msgObj.value}`;
       }
       if (issuesArray[x].type === "EMPTY") {
-        issuesMsg = `Entity ${issuesArray[x].fieldName} has no instances. This could mean that the pool was created but did no transactions were detected on it.`;
+        issuesMsg = `Entity ${issuesArray[x].fieldName} has no instances. This could mean that the pool was created but no transactions were detected on it.`;
       }
       issuesMsgs.push(<li key={`${x}-${issuesArray[x].fieldName}`}>{issuesMsg}</li>);
     }
@@ -71,15 +70,26 @@ const messagesByLevel = (
 
 interface IssuesProps {
   issuesArrayProps: { message: string; type: string; level: string; fieldName: string }[];
+  allLoaded: boolean;
+  oneLoaded: boolean;
 }
 // The issues display function takes the issues object passed in and creates the elements/messages to be rendered
-export const IssuesDisplay = ({ issuesArrayProps }: IssuesProps) => {
+export const IssuesDisplay = ({ issuesArrayProps, allLoaded, oneLoaded }: IssuesProps) => {
   const [issuesArray, setIssuesArray] = useState<{ message: string; type: string; level: string; fieldName: string }[]>(
     [],
   );
   useEffect(() => {
     setIssuesArray(issuesArrayProps);
   }, [issuesArrayProps]);
+  let waitingElement: JSX.Element | null = (
+    <>
+      <Typography variant="h6">WAITING TO SCAN DATA FOR ISSUES...</Typography>
+      <CircularProgress sx={{ margin: 6 }} size={50} />
+    </>
+  );
+  if (!oneLoaded && issuesArray.length === 0) {
+    return <IssuesContainer $hasCritical={false}>{waitingElement}</IssuesContainer>;
+  }
 
   const criticalIssues = issuesArray.filter((iss) => iss.level === "critical");
   const errorIssues = issuesArray.filter((iss) => iss.level === "error");
@@ -128,6 +138,10 @@ export const IssuesDisplay = ({ issuesArrayProps }: IssuesProps) => {
     );
   }
 
+  if (allLoaded) {
+    waitingElement = null;
+  }
+
   if (issuesDisplayCount > 0) {
     return (
       <IssuesContainer $hasCritical={hasCritical}>
@@ -135,6 +149,7 @@ export const IssuesDisplay = ({ issuesArrayProps }: IssuesProps) => {
         {criticalElement}
         {errorElement}
         {warningElement}
+        {waitingElement}
       </IssuesContainer>
     );
   } else {
