+++ dashboard/src/interfaces/ProtocolDashboard.tsx
@@ -18,7 +18,6 @@ function ProtocolDashboard() {
   const tabString = searchParams.get("tab") || "";
   const poolIdString = searchParams.get("poolId") || "";
   const scrollToView = searchParams.get("view") || "";
-  const protocolIdString = searchParams.get("protocolId") || "";
   const skipAmtParam = Number(searchParams.get("skipAmt")) || 0;
 
   const navigate = useNavigate();
@@ -29,10 +28,8 @@ function ProtocolDashboard() {
       queryURL = subgraphParam;
     }
   }
-
   const [subgraphToQuery, setSubgraphToQuery] = useState({ url: queryURL, version: "" });
   const [poolId, setPoolId] = useState<string>(poolIdString);
-  const [protocolId, setprotocolId] = useState<string>(protocolIdString);
   const [skipAmt, paginate] = useState<number>(skipAmtParam);
 
   ChartJS.register(...registerables);
@@ -70,15 +67,22 @@ function ProtocolDashboard() {
   if (!schemaVersion && protocolSchemaData?.protocols[0].schemaVersion) {
     schemaVersion = protocolSchemaData?.protocols[0].schemaVersion;
   }
+  const protocolIdString = searchParams.get("protocolId") || protocolSchemaData?.protocols[0].id;
+  const [protocolId, setprotocolId] = useState<string>(protocolIdString);
 
   // The following section fetches the full data from the subgraph. It routes to query selection and then makes the request
   const {
     entitiesData,
-    entities,
     poolData,
     query: graphQuery,
     events,
     protocolFields,
+    financialsQuery,
+    dailyUsageQuery,
+    hourlyUsageQuery,
+    protocolTableQuery,
+    poolsQuery,
+    poolTimeseriesQuery,
   } = schema(protocolSchemaData?.protocols[0].type, schemaVersion);
 
   const queryMain = gql`
@@ -86,12 +90,79 @@ function ProtocolDashboard() {
   `;
   const [getData, { data, loading, error }] = useLazyQuery(queryMain, { variables: { poolId, protocolId }, client });
 
+  const [
+    getFinancialsData,
+    { data: financialsData, loading: financialsLoading, error: financialsError, refetch: financialsRefetch },
+  ] = useLazyQuery(
+    gql`
+      ${financialsQuery}
+    `,
+    { client },
+  );
+  const [
+    getDailyUsageData,
+    { data: dailyUsageData, loading: dailyUsageLoading, error: dailyUsageError, refetch: dailyUsageRefetch },
+  ] = useLazyQuery(
+    gql`
+      ${dailyUsageQuery}
+    `,
+    { client },
+  );
+  const [
+    getHourlyUsageData,
+    { data: hourlyUsageData, loading: hourlyUsageLoading, error: hourlyUsageError, refetch: hourlyUsageRefetch },
+  ] = useLazyQuery(
+    gql`
+      ${hourlyUsageQuery}
+    `,
+    { client },
+  );
+
+  const [getProtocolTableData, { data: protocolTableData, loading: protocolTableLoading, error: protocolTableError }] =
+    useLazyQuery(
+      gql`
+        ${protocolTableQuery}
+      `,
+      { client, variables: { protocolId: protocolIdString } },
+    );
+
+  const [
+    getPoolsListData,
+    { data: poolsListData, loading: poolListLoading, error: poolsListError, refetch: poolsListRefetch },
+  ] = useLazyQuery(
+    gql`
+      ${poolsQuery}
+    `,
+    { client },
+  );
+  const [
+    getPoolTimeseriesData,
+    {
+      data: poolTimeseriesData,
+      loading: poolTimeseriesLoading,
+      error: poolTimeseriesError,
+      refetch: poolTimeseriesRefetch,
+    },
+  ] = useLazyQuery(
+    gql`
+      ${poolTimeseriesQuery}
+    `,
+    { variables: { poolId }, client },
+  );
+
+  const queryPoolOverview = gql`
+    ${poolOverview(protocolSchemaData?.protocols[0].type, schemaVersion)}
+  `;
+
+  const [getPoolsOverviewData, { data: dataPools, error: poolOverviewError, loading: poolOverviewLoading }] =
+    useLazyQuery(queryPoolOverview, { client: client, variables: { skipAmt } });
+
   let tabNum = "1";
-  if (tabString.toUpperCase() === "POOL" || tabString.toUpperCase() === "MARKET") {
+  if (tabString.toUpperCase() === "POOLOVERVIEW") {
     tabNum = "2";
-  } else if (tabString.toUpperCase() === "EVENTS") {
+  } else if (tabString.toUpperCase() === "POOL") {
     tabNum = "3";
-  } else if (tabString.toUpperCase() === "POOLOVERVIEW") {
+  } else if (tabString.toUpperCase() === "EVENTS") {
     tabNum = "4";
   }
 
@@ -109,34 +180,20 @@ function ProtocolDashboard() {
     let skipAmtParam = "";
     let poolParam = "";
     if (newValue === "2") {
-      poolParam = `&poolId=${poolIdFromParam || poolId}`;
-      tabName = "pool";
-    } else if (newValue === "3") {
-      poolParam = `&poolId=${poolIdFromParam || poolId}`;
-      tabName = "events";
-    } else if (newValue === "4") {
       tabName = "poolOverview";
       if (skipAmt > 0) {
         skipAmtParam = `&skipAmt=${skipAmt}`;
       }
+    } else if (newValue === "3") {
+      poolParam = `&poolId=${poolIdFromParam || poolId}`;
+      tabName = "pool";
+    } else if (newValue === "4") {
+      poolParam = `&poolId=${poolIdFromParam || poolId}`;
+      tabName = "events";
     }
     navigate(`?endpoint=${subgraphParam}&tab=${tabName}${protocolParam}${poolParam}${skipAmtParam}`);
     setTabValue(newValue);
   };
-
-  const queryPoolOverview = gql`
-    ${poolOverview(protocolSchemaData?.protocols[0].type, schemaVersion)}
-  `;
-
-  const clientPoolOverview = useMemo(() => NewClient(subgraphToQuery.url), [subgraphToQuery.url]);
-  const {
-    data: dataPools,
-    error: poolOverviewError,
-    loading: poolOverviewLoading,
-  } = useQuery(queryPoolOverview, {
-    client: clientPoolOverview,
-    variables: { skipAmt },
-  });
 
   let pools: { [x: string]: any }[] = [];
   if (dataPools && data) {
@@ -147,8 +204,71 @@ function ProtocolDashboard() {
     // If the schema query request was successful, make the full data query
     if (protocolSchemaData) {
       getData();
+      getProtocolTableData();
     }
-  }, [protocolSchemaData, getData]);
+  }, [protocolSchemaData, getData, getProtocolTableData]);
+
+  useEffect(() => {
+    if (protocolTableData) {
+      getFinancialsData();
+    }
+  }, [protocolTableData, getFinancialsData]);
+
+  useEffect(() => {
+    if (financialsData) {
+      getDailyUsageData();
+    }
+  }, [financialsData, getDailyUsageData]);
+
+  useEffect(() => {
+    if (dailyUsageData) {
+      getHourlyUsageData();
+    }
+  }, [dailyUsageData, getHourlyUsageData]);
+
+  useEffect(() => {
+    if (poolId) {
+      getPoolTimeseriesData();
+    }
+  }, [poolId]);
+
+  useEffect(() => {
+    if (financialsError) {
+      financialsRefetch();
+    }
+  }, [financialsError]);
+
+  useEffect(() => {
+    if (dailyUsageError) {
+      dailyUsageRefetch();
+    }
+  }, [dailyUsageError]);
+
+  useEffect(() => {
+    if (hourlyUsageError) {
+      hourlyUsageRefetch();
+    }
+  }, [hourlyUsageError]);
+
+  useEffect(() => {
+    if (poolsListError) {
+      poolsListRefetch();
+    }
+  }, [poolsListError]);
+
+  useEffect(() => {
+    if (poolTimeseriesError) {
+      poolTimeseriesRefetch();
+    }
+  }, [poolTimeseriesError]);
+
+  useEffect(() => {
+    if (tabValue === "2") {
+      getPoolsOverviewData();
+    } else if (tabValue === "3" || tabValue === "4") {
+      getPoolsListData();
+    }
+  }, [tabValue, getPoolsOverviewData, getPoolsListData]);
 
   useEffect(() => {
     document.getElementById(scrollToView)?.scrollIntoView();
@@ -202,10 +322,12 @@ function ProtocolDashboard() {
       {!!data && (
         <AllDataTabs
           data={data}
-          entities={entities}
           entitiesData={entitiesData}
           tabValue={tabValue}
           protocolFields={protocolFields}
+          poolsListData={poolsListData}
+          poolListLoading={poolListLoading}
+          poolsListError={poolsListError}
           poolNames={PoolNames[data.protocols[0].type]}
           poolId={poolId}
           poolData={poolData}
@@ -218,6 +340,23 @@ function ProtocolDashboard() {
           skipAmt={skipAmt}
           setProtocolId={(x) => setprotocolId(x)}
           poolOverviewRequest={{ poolOverviewError, poolOverviewLoading }}
+          poolTimeseriesRequest={{ poolTimeseriesData, poolTimeseriesError, poolTimeseriesLoading }}
+          protocolTimeseriesData={{
+            financialsDailySnapshots: financialsData?.financialsDailySnapshots,
+            usageMetricsDailySnapshots: dailyUsageData?.usageMetricsDailySnapshots,
+            usageMetricsHourlySnapshots: hourlyUsageData?.usageMetricsHourlySnapshots,
+          }}
+          protocolTimeseriesLoading={{
+            financialsDailySnapshots: financialsLoading,
+            usageMetricsDailySnapshots: dailyUsageLoading,
+            usageMetricsHourlySnapshots: hourlyUsageLoading,
+          }}
+          protocolTimeseriesError={{
+            financialsDailySnapshots: financialsError,
+            usageMetricsDailySnapshots: dailyUsageError,
+            usageMetricsHourlySnapshots: hourlyUsageError,
+          }}
+          protocolTableData={protocolTableData}
         />
       )}
     </div>
