+++ dashboard/src/queries/yield/poolOverview.ts
@@ -20,7 +20,7 @@ export const schema = (version: string): string => {
 export const schema100 = (): string => {
   return `
     query Data($skipAmt: Int!) {
-        vaults(first: 100, skip: $skipAmt) {
+        vaults(first: 50, skip: $skipAmt) {
             id
             name
             fees{
@@ -50,7 +50,7 @@ export const schema100 = (): string => {
 export const schema110 = (): string => {
   return `
     query Data($skipAmt: Int!) {
-        vaults(first: 100, skip: $skipAmt) {
+        vaults(first: 50, skip: $skipAmt) {
             id
             name
             fees{
@@ -80,7 +80,7 @@ export const schema110 = (): string => {
 export const schema120 = (): string => {
   return `
     query Data($skipAmt: Int!) {
-        vaults(first: 100, skip: $skipAmt) {
+        vaults(first: 50, skip: $skipAmt) {
             id
         name
         symbol
